ai_services/requirements.txt
Flask
scikit-learn
pandas
joblib
SpeechRecognition
PyAudio
numpy
flask-cors
pytesseract 
pillow 
flask
pytesseract
opencv-python
CORS
gdown
cv2
flask
scikit-learn
pandas
numpy
joblib 
xgboost
imbalanced-learn
imblearn
lightgbm
catboost
sklearn

